feat: split @geit CLI and BASE logic
