+++ packages/goat-cli/bin/goat.js
