+++ packages/goat-cli/src/commands/build.js
