+++ packages/goat-cli/src/commands/init.js
