+++ packages/goat-cli/src/commands/project.js
