+++ packages/goat-cli/src/config/config.js
