+++ packages/goat-cli/src/config/goatConfig.js
