+++ packages/goat-cli/src/config/modules/updateConfig.js
