+++ packages/goat-cli/src/index.js
