+++ packages/goat-cli/src/methods/debug/timeFunction.js
