+++ packages/goat-cli/src/methods/helpers/confirm.js
