+++ packages/goat-cli/src/methods/init/init.js
