+++ packages/goat-cli/src/methods/init/partials/configFileName.js
