+++ packages/goat-cli/src/methods/init/partials/getPackageConfig.js
