+++ packages/goat-cli/src/methods/init/partials/initConfig.js
