+++ packages/goat-cli/src/methods/init/partials/packageFiles.js
