+++ packages/goat-cli/src/methods/init/partials/projectConfig.js
