+++ packages/goat-cli/src/methods/init/partials/questions.js
