+++ packages/goat-cli/src/methods/modules/addModule.js
