+++ packages/goat-cli/src/methods/modules/collectModules.js
