+++ packages/goat-cli/src/methods/modules/collectModules.test.js
