+++ packages/goat-cli/src/methods/modules/listModules.js
