+++ packages/goat-cli/src/methods/modules/loadModule.js
