+++ packages/goat-cli/src/methods/modules/manageModule.js
