+++ packages/goat-cli/src/methods/settings/getGlobalConfig.js
