+++ packages/goat-cli/src/methods/settings/getGlobalConfig.test.js
