+++ packages/goat-cli/src/methods/settings/template/.goat.settings.js
