+++ packages/goat-cli/src/methods/settings/updateGlobalSettings.js
