+++ packages/goat-cli/src/references/settings.js
