+++ packages/goat-cli/src/schemas/generateConfig.js
