+++ packages/goat-cli/src/schemas/schema.js
