+++ packages/goat-cli/src/schemas/writeConfig.js
