+++ packages/goat-cli/src/validators/schema.js
