+++ packages/goat/src/index.ts
@@ -1,0 +1,5 @@
+import Goat from './Goat';
+import goatEvents from './events/GoatEvents';
+import watch from './events/watch';
+
+export { Goat, watch, goatEvents };
